Resources with generateName instead of name
This commit makes it possible to have resource definitions
with generateName instead of name and refer to them by that
name in graph dependencies.

When generateName is used k8s appends random suffix to that name
so that unique object is produced each time. Thus when resources
with generateName are used in the graph, each deployment of the graph
produces new resources and is no longer idempotent