Stable/unstable flows
"Stable" flag was added to the the Flow class.
Stable flows reuse existing replicas (FlowDeployments) for that flow.
Thus if you run the same flow again it will use the same $AC_NAME
values and thus will try to create resources with the same names as
they were in the previous run. In other words stable = idempotent.

Flows that are not stable produce new replicas on each run and are
not idempotent. Sucn flows can be used to express actions that change
the application state like "add node to the cluster"