Support for multi-document YAMLs for example e2e tests
There is a code in e2e framework that creates resource definitions
and dependencies out of stored YAMLs in example folder.
This code wasn't ready for multi-document YAMLs (those with ---
delimiter). For such YAMLs onlt the first document in the file
was loaded. Since all the examples have such YAML for dependencies
it caused all the examples to deploy with wrong dependency graph which
defeats the whole purpose of the tests

This commit adds support for multi-document YAMLS in e2e framework