Flow aware graph cycle detection
Cycle detection algorithm used to detect cycles in the whole
dependency graph ignoring dependency labels and how they correspond
to flow selectors. In some cases this caused false positive result.
For example it could happen if there is a cycle between creation and
destruction flow paths even if they are not related and the deployment
succeeded.

This commit makes cycle detection flow aware. Now the cycles are
searched for each flow * 2 paths (construction, destruction).
The validation takes place for all known flows rather than for the one
which is we about to execute.