Use status cache for resource readiness probes
When deployment engine waits for resource to became ready,
it should use ScheduledResource.Status() rather than Status method of
the underlying resource. The difference is that for ScheduledResource,
the status may be cached or set externally. For example, this
happens when Wait() times out. Without this change, the goroutine
created by Wait() will not recognize it and will continue running and
querying resource status in endless loop until deployment ends